move all Common Components into UI
