+++ app-web/src/components/Cards/Card/Card.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faExternalLinkSquareAlt } from '@fortawesome/free-solid-svg-icons';
 import styles from './Card.module.css';
-import Link from '../../Common/Link';
+import Link from '../../UI/Link/Link';
 import { ARIA_LABEL_RESOURCE, ARIA_LABEL_REPO } from '../../../constants/strings';
 
 const Card = ({ sourcePath, sourceName, title, description, resourcePath }) => (
