+++ app-web/src/components/Cards/Cards.js
@@ -5,7 +5,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faGithub } from '@fortawesome/fontawesome-free-brands';
 import styles from './Cards.module.css';
 import Card from './Card/Card';
-import Link from '../Common/Link';
+import Link from '../UI/Link/Link';
 import Toggle from './Toggle';
 import { ARIA_LABEL_REPO } from '../../constants/strings';
 import { LARGE_SCREEN_LIMIT, SMALL_SCREEN_LIMIT } from '../../constants/ui';
