+++ app-web/src/components/PrimaryFilter/PrimaryFilter.js
@@ -23,7 +23,7 @@ import Aux from '../../hoc/auxillary';
 import * as actions from '../../store/actions/actions';
 import { MAIN_NAV_CONFIG } from '../../constants/ui';
 import { ARIA_LABEL_FILTER_SELECT } from '../../constants/strings';
-import Hamburger from '../Common/Hamburger';
+import Hamburger from '../UI/Hamburger/Hamburger';
 
 export const PrimaryFilter = ({
   filterSiphonNodes,
