+++ app-web/src/components/SourceNavigation/SourceNavigation.js
@@ -18,7 +18,7 @@ Created by Patrick Simonian
 import React, { Component } from 'react';
 import shortid from 'shortid';
 import PropTypes from 'prop-types';
-import Link from '../Common/Link';
+import Link from '../UI/Link/Link';
 import styles from './SourceNavigation.module.css';
 // navigation for dynamically created page components
 class SourceNavigation extends Component {
